requirements.txt
ffmpeg-python==0.2.0
future==0.18.3
imageio==2.31.2
numpy==1.25.2
Pillow==10.0.0
